backend/requirements.txt
Adafruit_Blinka==6.20.1
adafruit_circuitpython_dht==3.7.0
adafruit_dht==1.4.0
board==1.0
Flask==1.1.2
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
flask_sock==0.4.0
matplotlib==3.5.1
pyserial==3.5
simple_pid==1.0.1
